// app/lib/models/campaign.ts
export interface ICampaign {
    id: string;
    influencer: string;
    activity: string;
    imageUrl: string;
    progress: number;
    budget: number;
    perMillion: number;
    createdAt: string;
    maxEarnings: number;
    maxEarningsPerPost: number;
    maxSubmissions: number;
    minViewsForPayout: number;
    moneySpent: number;
    isActive: boolean;
    isPot:boolean;
    perMillionText: string;
    discordInvite: string;
    totalViews: number;
    totalLikes: number;
    totalComments: number;
    totalShares: number;
    endedAt?: string;
    updatedAt?: string;
}